refactor(desktop): drop the dead github card subtree
GithubCard had zero production consumers, the live PR path is
PrDetailPanel in GitHubStudio. knip missed it because test files
are graph entry points in knip.json, keeping the tree reachable.

Deleted 14 files (~1,006 non-test LOC, ~323 test LOC). Relocated
five live pieces the subtree still owned: Avatar and ReviewStateIcon
to features/github/components/, and formatDuration,
latestTerminalReviewsByAuthor, computeTabStatus to
features/github/utils/, updating every import site in GitHubStudio
and GithubIssueComments.
