+++ apps/desktop/src/features/github/components/Avatar.tsx
